fix(room-io): ownership-aware FrameProcessor lifecycle management
The base class _ParticipantInputStream managed processor lifecycle in
_close_stream() without understanding processor ownership, causing
direct FrameProcessors to be killed on every track transition (#5448).

Add _processor_owned flag and _update_processor() method to the base
class. Direct processors (externally provided) survive stream
transitions; selector-created processors are closed when replaced or
when the stream closes without replacement. aclose() closes
unconditionally regardless of ownership.

Merge room_io test files into a single tests/test_room_io.py.
