--- tests/test_room_io_teardown.py
@@ -1,143 +1,0 @@
-from __future__ import annotations
-
-import asyncio
-from collections import defaultdict
-from types import SimpleNamespace
-from unittest.mock import AsyncMock, MagicMock
-
-import pytest
-
-from livekit import rtc
-from livekit.agents.voice.room_io._input import _ParticipantInputStream
-from livekit.agents.voice.room_io._output import _ParticipantTranscriptionOutput
-from livekit.agents.voice.room_io.room_io import RoomIO
-
-
-class _FakeRoom:
-    def __init__(self) -> None:
-        self._events: dict[str, list[object]] = defaultdict(list)
-        self.remote_participants: dict[str, object] = {}
-        self.local_participant = SimpleNamespace(identity="local")
-        self.name = "test-room"
-
-    def on(self, event: str, callback: object) -> None:
-        self._events[event].append(callback)
-
-    def off(self, event: str, callback: object) -> None:
-        callbacks = self._events[event]
-        callbacks.remove(callback)
-        if not callbacks:
-            self._events.pop(event, None)
-
-    def listener_count(self, event: str) -> int:
-        return len(self._events.get(event, []))
-
-    def isconnected(self) -> bool:
-        return True
-
-    def register_text_stream_handler(self, topic: str, callback: object) -> None:
-        self.on(f"text:{topic}", callback)
-
-    def unregister_text_stream_handler(self, topic: str) -> None:
-        self._events.pop(f"text:{topic}", None)
-
-
-class _NoopAudioInputStream(_ParticipantInputStream[rtc.AudioFrame]):
-    def __init__(self, room: _FakeRoom) -> None:
-        super().__init__(room, track_source=rtc.TrackSource.SOURCE_MICROPHONE)
-
-    def _create_stream(
-        self, track: rtc.RemoteTrack, participant: rtc.Participant
-    ) -> rtc.AudioStream:
-        raise AssertionError("_create_stream should not be called in teardown tests")
-
-
-class _FakeWriter:
-    def __init__(self) -> None:
-        self.close_calls = 0
-
-    async def aclose(self, attributes: dict[str, str] | None = None) -> None:
-        self.close_calls += 1
-
-
-@pytest.mark.asyncio
-async def test_participant_input_stream_aclose_unregisters_track_unpublished() -> None:
-    room = _FakeRoom()
-    stream = _NoopAudioInputStream(room)
-
-    assert room.listener_count("track_subscribed") == 1
-    assert room.listener_count("track_unpublished") == 1
-    assert room.listener_count("token_refreshed") == 1
-
-    await stream.aclose()
-
-    assert room.listener_count("track_subscribed") == 0
-    assert room.listener_count("track_unpublished") == 0
-    assert room.listener_count("token_refreshed") == 0
-
-
-@pytest.mark.asyncio
-async def test_transcription_output_aclose_unregisters_and_closes_resources() -> None:
-    room = _FakeRoom()
-    output = _ParticipantTranscriptionOutput(room=room, participant=None)
-    legacy_output, stream_output = output._ParticipantTranscriptionOutput__outputs
-
-    legacy_output._flush_task = asyncio.create_task(asyncio.sleep(0))
-    writer = _FakeWriter()
-    stream_output._writer = writer
-
-    assert room.listener_count("track_published") == 2
-    assert room.listener_count("local_track_published") == 2
-
-    await output.aclose()
-    await output.aclose()
-
-    assert room.listener_count("track_published") == 0
-    assert room.listener_count("local_track_published") == 0
-    assert legacy_output._flush_task is not None and legacy_output._flush_task.done()
-    assert writer.close_calls == 1
-
-
-@pytest.mark.asyncio
-async def test_roomio_aclose_unregisters_disconnect_and_closes_transcription_outputs() -> None:
-    room = _FakeRoom()
-    agent_session = SimpleNamespace(
-        off=MagicMock(),
-        input=SimpleNamespace(audio=None, video=None),
-        output=SimpleNamespace(audio=None, transcription=None),
-    )
-    room_io = RoomIO(agent_session, room)
-
-    room.on("participant_connected", room_io._on_participant_connected)
-    room.on("connection_state_changed", room_io._on_connection_state_changed)
-    room.on("participant_disconnected", room_io._on_participant_disconnected)
-
-    order: list[str] = []
-
-    async def _mark(name: str) -> None:
-        order.append(name)
-
-    async def _close_sync() -> None:
-        await _mark("sync")
-
-    async def _close_user() -> None:
-        await _mark("user")
-
-    async def _close_agent() -> None:
-        await _mark("agent")
-
-    room_io._tr_synchronizer = SimpleNamespace(aclose=AsyncMock(side_effect=_close_sync))
-    room_io._user_tr_output = SimpleNamespace(aclose=AsyncMock(side_effect=_close_user))
-    room_io._agent_tr_output = SimpleNamespace(aclose=AsyncMock(side_effect=_close_agent))
-
-    assert room.listener_count("participant_disconnected") == 1
-
-    await room_io.aclose()
-
-    assert room.listener_count("participant_connected") == 0
-    assert room.listener_count("connection_state_changed") == 0
-    assert room.listener_count("participant_disconnected") == 0
-    assert order == ["sync", "user", "agent"]
-    room_io._tr_synchronizer.aclose.assert_awaited_once()
-    room_io._user_tr_output.aclose.assert_awaited_once()
-    room_io._agent_tr_output.aclose.assert_awaited_once()
